Add RTVCollector to share render targets

+++ src/passes/objectPicker.hpp
@@ -3,8 +3,8 @@
 #include <memory>
 
 #include <d3d11_4.h>
-#include <wrl.h>
 #include <glm/glm.hpp>
+#include <wrl.h>
 
 using namespace Microsoft::WRL;
 
@@ -16,7 +16,7 @@ class ObjectPicker
 {
 public:
 	ObjectPicker(ComPtr<ID3D11Device> device, ComPtr<ID3D11DeviceContext> context);
-	uint32_t readBackID;
+	float readBackID;
 	void dispatchPick(const ComPtr<ID3D11ShaderResourceView>& srv, const uint32_t* mousePos, Scene* scene);
 
 private:
@@ -28,5 +28,4 @@ private:
 	ComPtr<ID3D11Buffer> m_stagingBuffer;
 	ComPtr<ID3D11UnorderedAccessView> m_uav;
 	std::unique_ptr<ShaderManager> m_shaderManager;
-
 };
