+++ src/shaders/colorPicker.hlsl
@@ -6,13 +6,13 @@ cbuffer cbPickLocation : register(b0)
 }
 
 
-Texture2D<uint> gBufferObjectID : register(t0);
+Texture2D<float> gBufferObjectID : register(t0);
 
 RWStructuredBuffer<uint> pickedColor : register(u0);
 
 [numthreads(1, 1, 1)]
 void CS(uint3 DTid : SV_DISPATCHTHREADID)
 {
-	uint idColor = gBufferObjectID.Load(int3(mousePosX, mousePosY, 0));
+	float idColor = gBufferObjectID.Load(int3(mousePosX, mousePosY, 0));
 	pickedColor[0] = idColor;
 }
